requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==3.2.15
django-environ==0.11.2
djangorestframework==3.12.4
drf-yasg==1.21.7
exceptiongroup==1.2.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
smartypants==2.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
